backend/requirements.txt
numpy==1.24.3
Pillow==10.0.0
pydicom==2.4.3
python-gdcm==3.0.22
